i fucked up with how i did context so now i get to fix it with one big commit sorry
