+++ dozer/cogs/maintenance.py
@@ -5,7 +5,8 @@ import sys
 
 from discord.ext.commands import NotOwner
 
-from dozer.bot import DOZER_LOGGER, DozerContext
+from dozer.bot import DOZER_LOGGER
+from dozer.context import DozerContext
 from ._utils import *
